Refactor drawing heatmap into a separate script.
